+++ hydra-node/src/Hydra/Chain.hs
@@ -21,7 +21,7 @@ import Hydra.Cardano.Api (
  )
 import Hydra.Ledger (IsTx, TxIdType, UTxOType)
 import Hydra.Party (Party)
-import Hydra.Snapshot (ConfirmedSnapshot, Snapshot, SnapshotNumber)
+import Hydra.Snapshot (ConfirmedSnapshot, SnapshotNumber)
 
 -- | Contains the head's parameters as established in the initial transaction.
 data HeadParameters = HeadParameters
@@ -44,7 +44,7 @@ data PostChainTx tx
   | AbortTx {utxo :: UTxOType tx}
   | CollectComTx {utxo :: UTxOType tx}
   | CloseTx {confirmedSnapshot :: ConfirmedSnapshot tx}
-  | ContestTx {snapshot :: Snapshot tx}
+  | ContestTx {confirmedSnapshot :: ConfirmedSnapshot tx}
   | FanoutTx {utxo :: UTxOType tx}
   deriving stock (Generic)
 
@@ -83,7 +83,7 @@ data OnChainTx tx
   | OnAbortTx
   | OnCollectComTx
   | OnCloseTx {snapshotNumber :: SnapshotNumber}
-  | OnContestTx
+  | OnContestTx {snapshotNumber :: SnapshotNumber}
   | OnFanoutTx
   deriving (Generic)